Set fetching to false after store.merge
I think one of the reasons #448 might have been happening
is because the change event was emitted when fetching
was set at the top of the action handling code. In this way,
the listening component was updated while fetching was false
and the newly received data was being merged into the existing
data. The result of this would be a flash of content.

I noticed this was happening in the space store and the
app store so I set fetching to false in the merge callback.